feat: adicionado o enum, e refatorado a localização da pasta factories
