+++ src/domain/factories/tipo-pagamento.enum.ts
@@ -1,0 +1,4 @@
+export enum TipoPagamento {
+  PIX = 'pix',
+  BOLETO = 'boleto',
+}
